skeyes/requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
frozenlist==1.4.1
idna==3.10
multidict==6.1.0
# Editable install with no version control (opensky-api==1.3.0)
-e /home/ana_tatabitovska/skeyes/opensky-api/python
psycopg2-binary==2.9.9
py4j==0.10.9.7
pyspark==3.5.2
python_opensky==1.0.1
requests==2.32.3
urllib3==2.2.3
yarl==1.11.1
